+++ tests/testthat/test_fujisan.R
@@ -26,8 +26,5 @@ test_that("fujisan solver works as expected", {
                         1,2,5,3,3,5,3,2,5,1,0,0), nrow = 2, byrow = TRUE)
     s2 <- solve_fujisan(coins = puzzle2)
     g2 <- read_ppn(textConnection(s2$ppn))[[1]]
-    skip_on_ci()
-    expect_doppelganger("fujisan", function() {
-        pmap_piece(g2$dfs[[1]], envir = list(piecepack = pp_cfg()), default.units = "in")
-    })
+    verify_output("../text_diagrams/fujisan_solved.txt", cat_move(g2, move = "SetupFn."))
 })